\name{animation_ratings}
\alias{animation_ratings}

\title{
  Movie Ratings
}
\description{
  Ratings for a set of 2010 animation movies
}
\usage{
  animation_ratings
}
\format{
   A data frame with 55 observations on the following 6 variables.
  \describe{
  \item{userId}{user ID}
  \item{movieId}{movie ID}
  \item{rating}{numerical rating}
  \item{timestamp}{time when the rating was recorded}
  \item{title}{name of the movie}
  \item{Group_Number}{numerical ID of movie}
  }
}
\source{
MovieLens by GroupLens Research
}
\keyword{datasets}

