\name{arm_height}
\alias{arm_height}

\title{
  Arm span and height measurements
}
\description{
  Arm span and height measurements for a sample of students
}
\usage{
  arm_height
}
\format{
   A data frame with 20 observations on the following 2 variables.
  \describe{
  \item{arm}{length of arm span in cm}
  \item{height}{height in cm}
  }
}
\source{
Sample of college students
}
\keyword{datasets}

