\name{batting_2018}
\alias{batting_2018}

\title{
  Batting Statistics for 2018 Season
}
\description{
  Batting statistics collected for all players during the first month and remainder of 2018 baseball season}
\usage{
  batting_2018
}
\format{
   A data frame with 549 observations on the following 5 variables.
  \describe{
  \item{Name}{name of player}
  \item{AB.x}{number of at bats in first month}
  \item{H.x}{number of hits in first month}
  \item{AB.y}{number of at bats in remainder of season}
  \item{H.y}{number of hits in remainder of season}
  }
}
\source{
Data collected from Retrosheet.org.
}
\keyword{datasets}

