\name{BBS_survey}
\alias{BBS_survey}

\title{
  Trend Estimates of Bird Populations
}
\description{
  Trend Estimates for 28 Grassland Bird Spies
}
\usage{
  BBS_survey
}
\format{
   A data frame with 28 observations on the following 4 variables.
  \describe{
  \item{Species_Name}{name of bird species}
  \item{Trend}{trend estimate}
  \item{SE}{standard error of estimate}
  \item{N_Site}{number of observations at site}
  }
}
\source{
North American Breeding Bird Survey
}
\keyword{datasets}

