\name{book_stats}
\alias{book_stats}

\title{
  Text Statistics for Books
}
\description{
  Text statistics for a collection of books sold at Amazon.com}
\usage{
  book_stats
}
\format{
   A data frame with 21 observations on the following 3 variables.
  \describe{
  \item{Book}{name of book}
  \item{Complex.Words}{percentage of words in the book with three or more syllables}
  \item{Fog.Index}{number of years of formal education required to read and understand a passage of text}
  }
}
\source{
Data collected from Amazon.com website.
}
\keyword{datasets}

