\name{buffalo_jan}
\alias{buffalo_jan}

\title{
  Buffalo snowfall data
}
\description{
  Total snowfall in inches for 20 Januarys in Buffalo, New York
}
\usage{
  buffalo_jan
}
\format{
   A data frame with 20 observations on the following 2 variables.
  \describe{
  \item{SEASON}{Season}
  \item{JAN}{inches of total snowfall}
  }
}
\source{
National Weather Service, www.weather.gov
}
\keyword{datasets}

