\name{career_1978}
\alias{career_1978}

\title{
  Career Trajectory Data for Baseball Players
}
\description{
 Season on-base statistics for collection of MLB
 baseball players who were born in 1978}
\usage{
  career_1978
}
\format{
   A data frame with 399 observations on the following 6 variables.
  \describe{
  \item{nameLast}{last name of player}
  \item{Player}{id of player}
  \item{Age}{age of player}
  \item{AgeD}{deviation of age from 30}
  \item{PA}{number of plate appearances}
  \item{OB}{number of on-base events}
  }
}
\source{
Data collected from Lahman database.
}
\keyword{datasets}

