\name{CEsample}
\alias{CEsample}

\title{
  Expeditures of U.S. Households
}
\description{
  Expeditures of U.S. Households
}
\usage{
  CEsample
}
\format{
   A data frame with 1000 observations on the following 3 variables.
  \describe{
  \item{UrbanRural}{urban/rural status of CU - 1 = urban and 2 = rural}
  \item{TotalIncomeLastYear}{amount of CU income before taxes in the last 12 months}
  \item{TotalExpLastQ}{CU's total expenditure in the last quarter}
  }
}
\source{
U.S. Bureau of Labor Statistics
}
\keyword{datasets}

