\name{ComputerPriceSample}
\alias{ComputerPriceSample}

\title{
  Personal Computer Data
}
\description{
  Variables on a sample of personal computers
}
\usage{
  ComputerPriceSample
}
\format{
   A data frame with 500 observations on the following 5 variables.
  \describe{
  \item{Price}{sales price}
  \item{Speed}{clock speed in MHz}
  \item{HardDrive}{size of hard drive in MB}
  \item{Ram}{size of Ram in MB}
  \item{Premium}{premium status of manufacturer}
  }
}
\source{

}
\keyword{datasets}

