\name{Cowles}
\alias{Cowles}

\title{
  Personality and Volunteering
}
\description{
  Data from study to learn about personality determinants of volunteering
}
\usage{
  Cowles
}
\format{
   A data frame with 1421 observations on the following 5 variables.
  \describe{
  \item{subject}{subject number}
  \item{neuroticism}{measurement of neuroticism}
  \item{extraversion}{measurement of extraversion}
  \item{sex}{male or female}
  \item{volunteer}{no or yes}
  }
}
\source{
Unknown.
}
\keyword{datasets}

