\name{DeathHeartAttackDataNYCfull}
\alias{DeathHeartAttackDataNYCfull}

\title{
  Risk-adjusted mortality outcomes for all NYC hospitals
}
\description{
  Reported deaths from heart attack for hospitals in New York City
}
\usage{
  DeathHeartAttackDataNYCfull
}
\format{
   A data frame with 45 observations on the following 5 variables.
  \describe{
  \item{Borough}{borough in New York City}
  \item{Type}{type of hospital}
  \item{Cases}{number of heart attach cases}
  \item{Deaths}{number of deaths}
  }
}
\source{
New York State Department of Health
}
\keyword{datasets}

