\name{DeathHeartAttackManhattan}
\alias{DeathHeartAttackManhattan}

\title{
  Risk-adjusted mortality outcomes for Manhattan hospitals
}
\description{
  Reported deaths from heart attack for hospitals in Manhattan in New York City
}
\usage{
  DeathHeartAttackManhattan
}
\format{
   A data frame with 13 observations on the following 4 variables.
  \describe{
  \item{Hospital}{name of hospital}
  \item{Type}{type of hospital}
  \item{Cases}{number of heart attach cases}
  \item{Deaths}{number of deaths}
  }
}
\source{
New York State Department of Health
}
\keyword{datasets}

