\name{electricbills}
\alias{electricbills}

\title{
  Electricity Bills
}
\description{
  Electricity bills collected for all months for five years}
\usage{
  electricbills
}
\format{
   A data frame with 62 observations on the following 3 variables.
  \describe{
  \item{Year}{year}
  \item{Month}{number of month}
  \item{Amount}{electicity bill in dollars}
  }
}
\source{
Data collected for one household in Ohio
}
\keyword{datasets}

