\name{federalist_word_study}
\alias{federalist_word_study}

\title{
  Frequency use of words for Federalist Papers
}
\description{
  Frequency use of words for Federalist Papers written by either Alexander Hamilton or James Madison
}
\usage{
  federalist_word_study
}
\format{
   A data frame with 56853 observations on the following 7 variables.
  \describe{
  \item{Name}{name of Federalist paper}
  \item{Total}{total number of words}
  \item{word}{word that is counted}
  \item{N}{frequency of the word}
  \item{Rate}{fraction of words with that word}
  \item{Authorship}{author of paper}
  \item{Disputed}{is authorship disputed?}
  }
}
\source{

}
\keyword{datasets}

