\name{federer_time_to_serve}
\alias{federer_time_to_serve}

\title{
  Times to Serve for Roger Federer
}
\description{
  Measurements of time to serve for 20 serves of the tennis player Roger Federer
}
\usage{
  federer_time_to_serve
}
\format{
   A data frame with 20 observations on the following one variable.
  \describe{
  \item{time}{time to serve in seconds}
  }
}
\source{
Jeff Sackman
}
\keyword{datasets}

