\name{fire_calls}
\alias{fire_calls}

\title{
  Fire Calls for Zip Code Areas
}
\description{
  The number of fire calls and building fires for ten zip codes in Montgomery County, Pennsylvania
}
\usage{
  fire_calls
}
\format{
   A data frame with 10 observations on the following 3 variables.
  \describe{
  \item{Zip_Code}{zip code}
  \item{Fire_Calls}{number of fire calls}
  \item{Building_Fires}{number of building fires}
  }
}
\source{
kaggle.com
}
\keyword{datasets}

