\name{football_field_goals}
\alias{football_field_goals}

\title{
  Football Field Goals Dataset
}
\description{
  Field goal attempt data for three seasons of professional football}
\usage{
  football_field_goals
}
\format{
   A data frame with 3025 observations on the following 5 variables.
  \describe{
  \item{Team}{name of team}
  \item{Year}{football season}
  \item{Kicker}{last name of kicker}
  \item{Distance}{distance in feet of attempt}
  \item{Success}{attempt was successful (1) or not (0)}
  }
}
\source{
Data collected by Michael Lopez.
}
\keyword{datasets}

