\name{gas2017}
\alias{gas2017}

\title{
  Gas bill data
}
\description{
  Measurements of average temperature and natural gas bill for each month in 2017
}
\usage{
  gas2017
}
\format{
   A data frame with 12 observations on the following 3 variables.
  \describe{
  \item{Month}{abbreviation of month}
  \item{Temp}{average temperature}
  \item{Bill}{natural gas bill in dollars}
  }
}
\source{
Personal data collected by a homeowner in Ohio
}
\keyword{datasets}

