\name{GradSchoolAdmission}
\alias{GradSchoolAdmission}

\title{
  Graduate School Admission
}
\description{
  Study to see what variables are helpful in determining admission to Graduate School
}
\usage{
  GradSchoolAdmission
}
\format{
   A data frame with 400 observations on the following 3 variables.
  \describe{
  \item{Admission}{student was admitted (1) or not admitted (0)}
  \item{GRE}{GRE score}
  \item{GPA}{grade point average}
  }
}
\source{
Unknown.
}
\keyword{datasets}

