\name{house_prices}
\alias{house_prices}

\title{
  House price data
}
\description{
  Measurements of house size and selling price for a collection of homes in a city in Ohio
}
\usage{
  house_prices
}
\format{
   A data frame with 24 observations on the following 2 variables.
  \describe{
  \item{price}{selling price in $1000}
  \item{size}{square footage of house}
  }
}
\source{
Zillow.com
}
\keyword{datasets}

