\name{JAGS_hier_trajectory}
\alias{JAGS_hier_trajectory}

\title{
  JAGS Script for Hierarchical Career Trajectory
}
\description{
  Model script for JAGS to fit a hierarchical career trajectory model
}
\usage{
  JAGS_hier_trajectory()
}
\value{
A character string containing the model script
}
\keyword{datasets}

