\name{JAGS_regression_cond_means}
\alias{JAGS_regression_cond_means}

\title{
  JAGS Script for Regression Model using a Conditional Means prior
}
\description{
  Model script for JAGS to fit a simple linear regression model using a conditional means prior
}
\usage{
  JAGS_regression_cond_means()
}
\value{
A character string containing the model script
}
\keyword{datasets}

