\name{KDramaData}
\alias{KDramaData}

\title{
  Korean Drama Ratings
}
\description{
  Ratings of Korean dramas prodcast during different days of the week and didfferent producers
}
\usage{
  KDramaData
}
\format{
   A data frame with 101 observations on the following 5 variables.
  \describe{
  \item{Drama}{name of drama}
  \item{Schedule}{indicator of what day the drama was broadcast}
  \item{Producer}{indicator of the producer of the drama}
  \item{Rating}{rating of the drama}
  \item{Date}{date of rating}
  }
}
\source{
AGB Nielsen Media Research Group
}
\keyword{datasets}

