\name{LaborParticipation}
\alias{LaborParticipation}

\title{
  U.S. Women Labor Participation
}
\description{
  U.S. women labor participation and family income
}
\usage{
  LaborParticipation
}
\format{
   A data frame with 753 observations on the following 2 variables.
  \describe{
  \item{Participation}{labor participation of the wife}
  \item{FamilyIncome}{family income exclusive of wife's income in $1000}
  }
}
\source{
University of Michigan Panel Study of Income Dynamics
}
\keyword{datasets}

