\name{Madison_can}
\alias{Madison_can}

\title{
  Frequency use of "can" for Federalist Papers
}
\description{
  Frequency use of "can" for Federalist Papers written by James Madison
}
\usage{
  Madison_can
}
\format{
   A data frame with 49 observations on the following 6 variables.
  \describe{
  \item{Name}{name of Federalist paper}
  \item{Total}{total number of words}
  \item{word}{word that is counted}
  \item{N}{frequency of the word}
  \item{Rate}{fraction of words with that word}
  \item{Authorship}{author of paper}
  }
}
\source{
http://www.gutenberg.org/ebooks/18
}
\keyword{datasets}

