\name{marriage_counts}
\alias{marriage_counts}

\title{
  Annual Marriage Counts in Italy
}
\description{
  Annual marriage counts per 1000 of the population in Italy from 1936 to 1951
}
\usage{
  marriage_counts
}
\format{
   A data frame with 16 observations on the following 2 variables.
  \describe{
  \item{Year}{year}
  \item{Count}{count of marriages per 1000 people}
  }
}
\source{
Unknown.
}
\keyword{datasets}

