\name{mcdonalds}
\alias{mcdonalds}

\title{
  Nutritional data for McDonalds Sandwiches
}
\description{
 Serving size and calories for a selection of sandwiches from McDonalds
}
\usage{
  mcdonalds
}
\format{
   A data frame with 11 observations on the following 3 variables.
  \describe{
  \item{Sandwich}{name of sandwich}
  \item{Size}{serving size in grams}
  \item{Calories}{calories of sandwich}
  }
}
\source{
McDonalds restaurant
}
\keyword{datasets}

