\name{movies2017}
\alias{movies2017}

\title{
  Movies Sales Data
}
\description{
Weekend and gross sales for a selection of movies released in 2017
}
\usage{
  movies2017
}
\format{
   A data frame with 10 observations on the following 3 variables.
  \describe{
  \item{Movie}{name of movie}
  \item{Weekend}{opening weekend sales in millions of dollars}
  \item{Gross}{gross sales in millions of dollars}
  }
}
\source{
Internet Movie Database
}
\keyword{datasets}

