\name{nba_guards}
\alias{nba_guards}

\title{
  Basketball Shooting Data for Point Guards
}
\description{
 Field goal and free throw shooting data for a collection of great NBA point guards}
\usage{
  nba_guards
}
\format{
   A data frame with 230 observations on the following 6 variables.
  \describe{
  \item{Player}{name of player}
  \item{Age}{age of player}
  \item{FG}{field goals}
  \item{FGA}{field goal attempts}
  \item{FT}{free throws}
  \item{FTA}{free throw attempts}
  }
}
\source{
Data collected from Basketball-Reference.com.
}
\keyword{datasets}

