\name{olympic_butterfly}
\alias{olympic_butterfly}

\title{
 Winning Times in the 100 Meter Butterfly Race
}
\description{
Winning times in seconds for the men's and women's 100m butterfly race for the Olympics from 1964 through 2016.
}
\usage{
  olympic_butterfly
}
\format{
   A data frame with 28 observations on the following 3 variables.
  \describe{
  \item{Year}{year of Olympics}
  \item{Gender}{gender}
  \item{Time}{winning time in seconds}
  }
}
\source{
https://www.olympic.org/swimming/
}
\keyword{datasets}

