\name{ProfessorSalary}
\alias{ProfessorSalary}

\title{
  Professor Salary Study
}
\description{
  Study on inputs that impact a salary of a professor
}
\usage{
  ProfessorSalary
}
\format{
   A data frame with 397 observations on the following 7 variables.
  \describe{
  \item{subject}{subject id}
  \item{rank}{professor rank}
  \item{discipline}{A is theoretical and B is applied}
  \item{yrs.since.phd}{number of years since receipt of doctorate}
  \item{yrs.service}{number of years of service}
  \item{sex}{Female or Male}
  \item{salary}{nine-month salary in dollars}
  }
}
\source{
Unknown.
}
\keyword{datasets}

