\name{pt100price}
\alias{pt100price}

\title{
 Prices of One Carat Diamonds
}
\description{
Prices of a sample of one carat diamonds
}
\usage{
  pt100price
}
\format{
   A data frame with 25 observations on the following 2 variables.
  \describe{
  \item{diamond}{index of diamond}
  \item{price}{price divided by 100}
  }
}
\source{
Unknown.
}
\keyword{datasets}

