\name{pt99price}
\alias{pt99price}

\title{
 Prices of 0.99 Carat Diamonds
}
\description{
Prices of a sample of 0.99 carat diamonds
}
\usage{
  pt99price
}
\format{
   A data frame with 23 observations on the following 2 variables.
  \describe{
  \item{diamond}{index of diamond}
  \item{price}{price divided by 100}
  }
}
\source{
Unknown.
}
\keyword{datasets}

