\name{pythag2018}
\alias{pythag2018}

\title{
 Baseball Win-Loss Records
}
\description{
Final standings of the MLB baseball teams in the 2018 season
}
\usage{
  pythag2018
}
\format{
   A data frame with 30 observations on the following 7 variables.
  \describe{
  \item{Team}{team abbreviation}
  \item{League}{league abbreviation}
  \item{W}{number of wins}
  \item{L}{number of losses}
  \item{Pct}{proportion of wins}
  \item{R}{average runs scored}
  \item{RA}{average runs allowed}
  }
}
\source{
Lahman database
}
\keyword{datasets}

