\name{ScoreData}
\alias{ScoreData}

\title{
  Scores on Achievement Exam
}
\description{
  Scores on a 20-question T/F exam
}
\usage{
  ScoreData
}
\format{
   A data frame with 30 observations on the following 2 variables.
  \describe{
  \item{Person}{subject id}
  \item{Score}{number correct in 20-question exam}
  }
}
\source{
Data randomly generated.
}
\keyword{datasets}

