\name{sleeping_times}
\alias{sleeping_times}

\title{
 Sleeping Times
}
\description{
Sample of sleeping times for a single night for a sample of college students
}
\usage{
  sleeping_times
}
\format{
   A data frame with 14 observations on the following single variable.
  \describe{
  \item{hours}{number of hours of sleep}
  }
}
\source{
Personal collection
}
\keyword{datasets}

