\name{taxi_fares}
\alias{taxi_fares}

\title{
 Taxi Fares
}
\description{
Sample of taxi fares from a particular city
}
\usage{
  taxi_fares
}
\format{
   A data frame with 20 observations on the following single variable.
  \describe{
  \item{fare}{taxi cab fare}
  }
}
\source{
Personal collection
}
\keyword{datasets}

