\name{tennis_serve}
\alias{tennis_serve}

\title{
 Tennis Times to Serve
}
\description{
Data on time to serve for six professional tennis players
}
\usage{
  tennis_serve
}
\format{
   A data frame with 6 observations on the following 3 variables.
  \describe{
  \item{Player}{last name of player}
  \item{n}{number of serves}
  \item{ybar}{mean time to serve}
  }
}
\source{
https://github.com/JeffSackmann
}
\keyword{datasets}

