\name{trout20}
\alias{trout20}

\title{
 Mike Trout Statcast Data
}
\description{
Launch speed and distance traveled for a sample of balls hit by the baseball player Mike Trout
}
\usage{
  trout20
}
\format{
   A data frame with 25 observations on the following 2 variables.
  \describe{
  \item{launch_speed}{launch speed in mph}
  \item{hit_distance_sc}{distance in feet}
  }
}
\source{
Major League Baseball Advanced Media
}
\keyword{datasets}

