\name{two_players_time_to_serve}
\alias{two_players_time_to_serve}

\title{
  Times to Serve for Two Tennis Players
}
\description{
  Measurements of time to serve serves of the tennis players Roger Federer and Rafael Nadal
}
\usage{
  two_players_time_to_serve
}
\format{
   A data frame with 100 observations on the following 2 variables.
  \describe{
  \item{Player}{last name of player}
  \item{time}{time to serve in seconds}
  }
}
\source{
Jeff Sackman
}
\keyword{datasets}

