\name{web_visits}
\alias{web_visits}

\title{
 Website tracking data
}
\description{
Number of visits to a blog website for different weeks and days of the week
}
\usage{
  web_visits
}
\format{
   A data frame with 28 observations on the following 3 variables.
  \describe{
  \item{Week}{week number}
  \item{Day}{day ofthe week}
  \item{Count}{number of website visits}
  }
}
\source{
Personal data collected from Wordpress.com
}
\keyword{datasets}

