\name{HWhours5schools}
\alias{HWhours5schools}

\title{
  Homework Hours for Five Schools
}
\description{
  Weekly hours spent on homework for students from five schools
}
\usage{
  HWhours5schools
}
\format{
   A data frame with 116 observations on the following 2 variables.
  \describe{
  \item{school}{school number of student}
  \item{score}{weekly hours spent on homework}
  }
}
\source{

}
\keyword{datasets}

